+++ src/ResourceManager/Network/Commands.Network/ApplicationGateway/Probe/NewAzureApplicationGatewayProbeConfigCommand.cs
@@ -17,7 +17,7 @@ using System.Management.Automation;
 
 namespace Microsoft.Azure.Commands.Network
 {
-    [Cmdlet(VerbsCommon.New, "AzureRmApplicationGatewayProbeConfig"), OutputType(typeof(PSApplicationGatewayProbe))]
+    [Cmdlet("New", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "ApplicationGatewayProbeConfig"), OutputType(typeof(PSApplicationGatewayProbe))]
     public class NewAzureApplicationGatewayProbeConfigCommand : AzureApplicationGatewayProbeConfigBase
     {
         public override void ExecuteCmdlet()