+++ src/ResourceManager/Maps/Commands.Maps/ChangeLog.md
@@ -18,6 +18,10 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Fixed issue with default resource groups not being set.
+
+## Version 1.0.3
+* Updated to the latest version of the Azure ClientRuntime.
 
 ## Version 1.0.2
 * Updated help files to include full parameter types and correct input/output types.